feature(condo): DOMA-8413 remove test duplicate
