+++ apps/condo/domains/billing/schema/RegisterBillingReceiptsService.test.js
@@ -1,1028 +1,400 @@
 /**
  * Generated by `createservice billing.RegisterBillingReceiptsService --type mutations`
  */
-const { faker } = require('@faker-js/faker')
+const { faker  } = require('@faker-js/faker/locale/ru')
+const Big = require('big.js')
+const dayjs = require('dayjs')
 
 const {
     catchErrorFrom,
-    expectToThrowGQLError,
 } = require('@open-condo/keystone/test.utils')
-const { expectToThrowAuthenticationError, expectToThrowAccessDeniedErrorToResult } = require('@open-condo/keystone/test.utils')
-const { makeClient, makeLoggedInAdminClient } = require('@open-condo/keystone/test.utils')
+const { makeLoggedInAdminClient } = require('@open-condo/keystone/test.utils')
+const { expectToThrowAccessDeniedErrorToResult } = require('@open-condo/keystone/test.utils')
 
-const { DEFAULT_BILLING_CATEGORY_ID } = require('@condo/domains/billing/constants/constants')
-const { ERRORS: mutationErrors } = require('@condo/domains/billing/constants/registerBillingReceiptService')
+const { BILLING_ACCOUNT_OWNER_TYPE_COMPANY, BILLING_ACCOUNT_OWNER_TYPE_PERSON } = require('@condo/domains/billing/constants/constants')
+const {
+    ERRORS,
+} = require('@condo/domains/billing/constants/registerBillingReceiptService')
 const { registerBillingReceiptsByTestClient } = require('@condo/domains/billing/utils/testSchema')
 const {
-    makeServiceUserForIntegration,
-    makeOrganizationIntegrationManager,
-    createTestBillingProperty,
-    createTestBillingAccount,
     createTestBillingIntegration,
     createTestBillingIntegrationOrganizationContext,
     BillingReceipt,
     BillingAccount,
-    BillingCategory,
-    BillingProperty,
+    BillingRecipient,
     generateServicesData,
-    updateTestBillingIntegrationAccessRight,
-    createRegisterBillingReceiptsPayload,
 } = require('@condo/domains/billing/utils/testSchema')
-const { createTestBillingCategory } = require('@condo/domains/billing/utils/testSchema')
+const { createTestBillingCategory, createTestBillingIntegrationAccessRight } = require('@condo/domains/billing/utils/testSchema')
 const { createTestOrganization } = require('@condo/domains/organization/utils/testSchema')
-const { FLAT_UNIT_TYPE, APARTMENT_UNIT_TYPE } = require('@condo/domains/property/constants/common')
-const {
-    makeClientWithSupportUser,
-    makeLoggedInClient,
-} = require('@condo/domains/user/utils/testSchema')
+const { makeClientWithServiceUser } = require('@condo/domains/user/utils/testSchema')
 
+const createAddressWithUnit = () => `${faker.address.cityName()} ${faker.address.streetAddress(true)}`
+
+const createJSONReceipt = (extra = {}) => {
+    const [month, year] = dayjs().add(-1, 'month').format('MM-YYYY').split('-').map(Number)
+    return Object.fromEntries(Object.entries({
+        importId: faker.datatype.uuid(),
+        address: createAddressWithUnit(),
+        accountNumber: randomNumber(10).toString(),
+        toPay: faker.finance.amount(-100, 5000),
+        month,
+        year,
+        services: generateServicesData(faker.datatype.number({ min: 3, max: 5 })),
+        ...createRecipient(),
+        raw: extra,
+        ...extra,
+    }).filter(([, value]) => !!value))
+}
+
+const createRecipient = (extra = {}) => {
+    return {
+        tin: faker.random.numeric(8),
+        routingNumber: faker.random.numeric(5),
+        bankAccount: faker.random.numeric(12),
+        ...extra,
+    }
+}
+
+const createValidELS = () => `${randomNumber(2)}БГ${randomNumber(6)}`
+
+const randomNumber = (numDigits) => {
+    const min = 10 ** (numDigits - 1)
+    const max = 10 ** numDigits - 1
+    return faker.datatype.number({ min, max })
+}
 
 describe('RegisterBillingReceiptsService', () => {
-    let admin
-    let support
-    let anonymous
-    let user
+
+    const clients = {}
+    let integration = {}
 
     beforeAll(async () => {
-        admin = await makeLoggedInAdminClient()
-        support = await makeClientWithSupportUser()
-        anonymous = await makeClient()
-        user = await makeLoggedInClient()
+        clients.admin = await makeLoggedInAdminClient()
+        const [organization] = await createTestOrganization(clients.admin)
+        const [billingIntegration] = await createTestBillingIntegration(clients.admin)
+        clients.serviceUserWithoutAccessRights = await makeClientWithServiceUser()
+        clients.serviceUserWithAccessRights = await makeClientWithServiceUser()
+        await createTestBillingIntegrationAccessRight(clients.admin, billingIntegration, clients.serviceUserWithAccessRights['user'])
+        const [billingContext] = await createTestBillingIntegrationOrganizationContext(clients.admin, organization, billingIntegration)
+        integration = { organization, billingIntegration, billingContext }
         process.env.ADDRESS_SERVICE_CLIENT_MODE = 'fake'
     })
 
-    describe('Execute', () => {
-        test('Admin can execute mutation', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const payload = {
-                context: { id: billingContext.id },
+    describe('Access rights', () => {
+        test('Admin can execute RegisterBillingReceiptsService', async () => {
+            const [receipts] = await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
                 receipts: [],
-            }
-
-            const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-
-            expect(data).toHaveLength(0)
+            })
+            expect(receipts).toHaveLength(0)
         })
-
-        test('Billing service account can execute mutation', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const integrationUser = await makeServiceUserForIntegration(integration)
-
-            const payload = {
-                context: { id: billingContext.id },
+        test('Service account can execute RegisterBillingReceiptsService', async () => {
+            const [receipts] = await registerBillingReceiptsByTestClient(clients.serviceUserWithAccessRights, {
+                context: { id: integration.billingContext.id },
                 receipts: [],
-            }
-
-            const [ data ] = await registerBillingReceiptsByTestClient(integrationUser, payload)
-
-            expect(data).toHaveLength(0)
+            })
+            expect(receipts).toHaveLength(0)
         })
-
-        test('Billing service account without Access Right cant execute mutation', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const integrationUser = await makeServiceUserForIntegration(integration)
-            await updateTestBillingIntegrationAccessRight(admin, integrationUser.accessRight.id, { deletedAt: 'true' })
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [],
-            }
-
+        test('Service account without Access Right can not execute RegisterBillingReceiptsService', async () => {
             await expectToThrowAccessDeniedErrorToResult(async () => {
-                await registerBillingReceiptsByTestClient(integrationUser, payload)
-            })
-        })
-
-        test('Organization employee can not execute mutation', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-            const { managerUserClient } = await makeOrganizationIntegrationManager({ context: billingContext })
-            await expectToThrowAccessDeniedErrorToResult(async () => {
-                await registerBillingReceiptsByTestClient(managerUserClient, { context: { id: '1234' }, receipts: [] })
-            })
-        })
-
-        test('Support can not execute mutation', async () => {
-            await expectToThrowAccessDeniedErrorToResult(async () => {
-                await registerBillingReceiptsByTestClient(support, { context: { id: '1234' }, receipts: [] })
-            })
-        })
-
-        test('User can not execute mutation', async () => {
-            await expectToThrowAccessDeniedErrorToResult(async () => {
-                await registerBillingReceiptsByTestClient(user, { context: { id: '1234' }, receipts: [] })
-            })
-        })
-
-        test('Anonymous can not execute mutation', async () => {
-            await expectToThrowAuthenticationError(async () => {
-                await registerBillingReceiptsByTestClient(anonymous, { context: { id: '1234' }, receipts: [] })
-            }, 'result')
-        })
-    })
-
-    describe('Business Logic', () => {
-
-        describe('BillingProperties', () => {
-            test('BillingProperties are created and not deleted', async () => {
-                /**
-                 * Existing properties = {p1, p2}
-                 * Properties from args = {p1, p3}
-                 * Result: {p1, p2, p3}
-                 */
-
-                const EXISTING_TEST_ADDRESS_P1 = 'г. Екатеринбург, Тургенева 4'
-                const EXISTING_TEST_ADDRESS_P2 = 'г. Екатеринбург, Проспект Ленина, 44'
-                const NON_EXISTING_TEST_ADDRESS_P3 = 'г. Екатеринбург, Проспект Космонавтов 11б'
-
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                await createTestBillingProperty(admin, billingContext, {
-                    address: EXISTING_TEST_ADDRESS_P1,
+                await registerBillingReceiptsByTestClient(clients.serviceUserWithoutAccessRights, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [],
                 })
-
-                await createTestBillingProperty(admin, billingContext, {
-                    address: EXISTING_TEST_ADDRESS_P2,
-                })
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        createRegisterBillingReceiptsPayload({
-                            address: EXISTING_TEST_ADDRESS_P1,
-                        }),
-                        createRegisterBillingReceiptsPayload({
-                            address: NON_EXISTING_TEST_ADDRESS_P3,
-                        }),
-                    ],
-                }
-
-                const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingAccounts = await BillingAccount.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-                expect(billingProperties).toHaveLength(3)
-                expect(billingAccounts).toHaveLength(2)
-                expect(billingReceipts).toHaveLength(2)
-                expect(data).toHaveLength(2)
-            })
-        })
-
-        describe('BillingAccounts', () => {
-            test('BillingAccounts are created and not deleted', async () => {
-                /**
-                 * Existing properties = {p1}
-                 * Properties from args = {p2}
-                 * Existing accounts = {a1, a2} for p1
-                 * Accounts from args = {a1, a3, a1', a2', a3'} // a1' a2' a3' are for p2
-                 *
-                 * Result Accounts: p1 -> {a1,  a2,  a3 }
-                 *                  p2 -> {a1', a2', a3'}
-                 */
-
-                const TEST_ADDRESS_P1 = 'г. Екатеринбург, Тургенева 4'
-                const TEST_ADDRESS_P2 = 'г. Екатеринбург, Проспект Космонавтов 11б'
-
-                const TEST_BILLING_ACCOUNT_1_UNITNAME = '1'
-                const TEST_BILLING_ACCOUNT_1_NUMBER = 'n1'
-
-                const TEST_BILLING_ACCOUNT_2_UNITNAME = '2'
-                const TEST_BILLING_ACCOUNT_2_NUMBER = 'n2'
-
-                const TEST_BILLING_ACCOUNT_3_UNITNAME = '3'
-                const TEST_BILLING_ACCOUNT_3_NUMBER = 'n3'
-
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const [billingProperty1] = await createTestBillingProperty(admin, billingContext, {
-                    address: TEST_ADDRESS_P1,
-                })
-
-                await createTestBillingAccount(admin, billingContext, billingProperty1, {
-                    number: TEST_BILLING_ACCOUNT_1_NUMBER,
-                    unitName: TEST_BILLING_ACCOUNT_1_UNITNAME,
-                    unitType: FLAT_UNIT_TYPE,
-                })
-
-                await createTestBillingAccount(admin, billingContext, billingProperty1, {
-                    number: TEST_BILLING_ACCOUNT_2_NUMBER,
-                    unitName: TEST_BILLING_ACCOUNT_2_UNITNAME,
-                    unitType: FLAT_UNIT_TYPE,
-                })
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        createRegisterBillingReceiptsPayload({
-                            address: TEST_ADDRESS_P1,
-
-                            unitType: FLAT_UNIT_TYPE,
-                            accountNumber: TEST_BILLING_ACCOUNT_1_NUMBER,
-                            unitName: TEST_BILLING_ACCOUNT_1_UNITNAME,
-                        }),
-                        createRegisterBillingReceiptsPayload({
-                            address: TEST_ADDRESS_P1,
-
-                            unitType: FLAT_UNIT_TYPE,
-                            accountNumber: TEST_BILLING_ACCOUNT_3_NUMBER,
-                            unitName: TEST_BILLING_ACCOUNT_3_UNITNAME,
-                        }),
-                        createRegisterBillingReceiptsPayload({
-                            address: TEST_ADDRESS_P2,
-
-                            unitType: FLAT_UNIT_TYPE,
-                            accountNumber: TEST_BILLING_ACCOUNT_1_NUMBER,
-                            unitName: TEST_BILLING_ACCOUNT_1_UNITNAME,
-                        }),
-                        createRegisterBillingReceiptsPayload({
-                            address: TEST_ADDRESS_P2,
-
-                            unitType: FLAT_UNIT_TYPE,
-                            accountNumber: TEST_BILLING_ACCOUNT_2_NUMBER,
-                            unitName: TEST_BILLING_ACCOUNT_2_UNITNAME,
-                        }),
-                        createRegisterBillingReceiptsPayload({
-                            address: TEST_ADDRESS_P2,
-
-                            unitType: FLAT_UNIT_TYPE,
-                            accountNumber: TEST_BILLING_ACCOUNT_3_NUMBER,
-                            unitName: TEST_BILLING_ACCOUNT_3_UNITNAME,
-                        }),
-                    ],
-                }
-
-                const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingAccounts = await BillingAccount.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(2)
-                expect(billingAccounts).toHaveLength(6)
-                expect(billingReceipts).toHaveLength(5)
-                expect(data).toHaveLength(5)
-            })
-
-            test('BillingAccounts are created if they have different UnitTypes', async () => {
-
-                const TEST_ADDRESS_P1 = 'г. Екатеринбург, Тургенева 4'
-
-                const TEST_BILLING_ACCOUNT_1_UNITNAME = '1'
-                const TEST_BILLING_ACCOUNT_1_NUMBER = 'n1'
-
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        createRegisterBillingReceiptsPayload({
-                            address: TEST_ADDRESS_P1,
-
-                            unitType: FLAT_UNIT_TYPE,
-                            accountNumber: TEST_BILLING_ACCOUNT_1_NUMBER,
-                            unitName: TEST_BILLING_ACCOUNT_1_UNITNAME,
-
-                            month: 4,
-                        }),
-                        createRegisterBillingReceiptsPayload({
-                            address: TEST_ADDRESS_P1,
-
-                            unitType: APARTMENT_UNIT_TYPE,
-                            accountNumber: TEST_BILLING_ACCOUNT_1_NUMBER,
-                            unitName: TEST_BILLING_ACCOUNT_1_UNITNAME,
-
-                            month: 4,
-                        }),
-                        createRegisterBillingReceiptsPayload({
-                            address: TEST_ADDRESS_P1,
-
-                            unitType: APARTMENT_UNIT_TYPE,
-                            accountNumber: TEST_BILLING_ACCOUNT_1_NUMBER,
-                            unitName: TEST_BILLING_ACCOUNT_1_UNITNAME,
-
-                            month: 5,
-                        }),
-                    ],
-                }
-
-                const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingAccounts = await BillingAccount.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(1)
-                expect(billingAccounts).toHaveLength(2)
-                expect(billingReceipts).toHaveLength(3)
-                expect(data).toHaveLength(3)
-            })
-        })
-
-        describe('BillingReceipts', () => {
-            test('BillingReceipts are created if they are in different addresses', async () => {
-                const TEST_ADDRESS_P1 = 'г. Екатеринбург, Тургенева 4'
-                const TEST_ADDRESS_P2 = 'г. Екатеринбург, Проспект Ленина, 44'
-
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const receiptInput = createRegisterBillingReceiptsPayload()
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            address: TEST_ADDRESS_P1,
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                        {
-                            ...receiptInput,
-                            address: TEST_ADDRESS_P2,
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                    ],
-                }
-
-                const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(2)
-                expect(billingReceipts).toHaveLength(2)
-                expect(data).toHaveLength(2)
-            })
-
-            test('BillingReceipts are created if they have different accounts', async () => {
-
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const receiptInput = createRegisterBillingReceiptsPayload()
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            accountNumber: '1',
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                        {
-                            ...receiptInput,
-                            accountNumber: '2',
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                    ],
-                }
-
-                const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(1)
-                expect(billingReceipts).toHaveLength(2)
-                expect(data).toHaveLength(2)
-            })
-
-            test('BillingReceipts are created if they have different categories', async () => {
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                // create categories
-                const services = generateServicesData()
-                const categories = (await BillingCategory.getAll(admin, {}))
-                    .filter(category => category.id !== DEFAULT_BILLING_CATEGORY_ID)
-                const billingCategoryWithServices = (await createTestBillingCategory(
-                    admin, { name: `Category ${new Date()}`, serviceNames: services.map(({ name }) => name) },
-                ))[0]
-
-                const receiptInput = createRegisterBillingReceiptsPayload()
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            category: { id: categories[0].id },
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                        {
-                            ...receiptInput,
-                            category: { id: categories[1].id },
-                            importId: faker.random.alphaNumeric(24),
-                        },
-
-                        // include 3 receipts with the same category services
-                        // (going to be merged into one receipt since it related to the same category)
-                        {
-                            ...receiptInput,
-                            services: services.slice(0, 1),
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                        {
-                            ...receiptInput,
-                            services: services.slice(0, 2),
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                        {
-                            ...receiptInput,
-                            services,
-                            importId: faker.random.alphaNumeric(24),
-                        },
-
-                        // should fall back to default category
-                        // for no services case
-                        {
-                            ...receiptInput,
-                            services: [],
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                    ],
-                }
-
-                delete payload.receipts[2].category
-                delete payload.receipts[3].category
-                delete payload.receipts[4].category
-                delete payload.receipts[5].category
-
-                const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-                const serviceCreatedReceipts = billingReceipts.filter(
-                    ({ category: { id } }) => id === billingCategoryWithServices.id
-                )
-                const receiptsWithNotCategoryInfo = billingReceipts.filter(({ importId }) => importId === payload.receipts[5].importId)
-
-                expect(billingProperties).toHaveLength(1)
-                expect(billingReceipts).toHaveLength(4)
-                expect(serviceCreatedReceipts).toHaveLength(1)
-                expect(receiptsWithNotCategoryInfo).toHaveLength(1)
-                expect(receiptsWithNotCategoryInfo[0].category.id).toEqual(DEFAULT_BILLING_CATEGORY_ID)
-                expect(data).toHaveLength(4)
-            })
-
-            test('BillingReceipts are created if they have different periods', async () => {
-
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const receiptInput = createRegisterBillingReceiptsPayload()
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            month: 3,
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                        {
-                            ...receiptInput,
-                            month: 10,
-                            importId: faker.random.alphaNumeric(24),
-                        },
-                    ],
-                }
-
-                const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(1)
-                expect(billingReceipts).toHaveLength(2)
-                expect(data).toHaveLength(2)
-            })
-
-            test('BillingReceipts are not changed if they refer to the same receipt', async () => {
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const receiptInput = createRegisterBillingReceiptsPayload()
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        receiptInput,
-                    ],
-                }
-                await registerBillingReceiptsByTestClient(admin, payload)
-
-                const existingBillingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-                const existingBillingReceiptId = existingBillingReceipts[0].id
-                const existingBillingReceiptUpdatedAt = existingBillingReceipts[0].updatedAt
-
-                const [data] = await registerBillingReceiptsByTestClient(admin, payload)
-
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingAccounts = await BillingAccount.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(1)
-                expect(billingAccounts).toHaveLength(1)
-                expect(billingReceipts).toHaveLength(1)
-                expect(data).toHaveLength(0)
-                expect(billingReceipts[0].id).toEqual(existingBillingReceiptId)
-                expect(billingReceipts[0].updatedAt).toEqual(existingBillingReceiptUpdatedAt)
-            })
-
-            test('BillingReceipts are updated if only toPay is different', async () => {
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const receiptInput = createRegisterBillingReceiptsPayload()
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            toPay: '190.00',
-                        },
-                    ],
-                }
-                const [ existingBillingReceipts ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const existingBillingReceiptId = existingBillingReceipts[0].id
-
-                const payload2 = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            toPay: '200.00',
-                        },
-                    ],
-                }
-                const [data] = await registerBillingReceiptsByTestClient(admin, payload2)
-
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(1)
-                expect(billingReceipts).toHaveLength(1)
-                expect(data).toHaveLength(1)
-                expect(billingReceipts[0].id).toEqual(existingBillingReceiptId)
-                expect(billingReceipts[0].toPay).toEqual('200.00000000')
-            })
-
-            test('BillingReceipts are updated if only toPayDetails are different', async () => {
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const receiptInput = createRegisterBillingReceiptsPayload()
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            toPayDetails: { 'formula':'balance + charge + recalculation + privilege + penalty', 'charge':'100.00', 'balance':'0.00', 'recalculation':null, 'privilege':null, 'penalty':null },
-                        },
-                    ],
-                }
-                const [ existingBillingReceipts ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const existingBillingReceiptId = existingBillingReceipts[0].id
-
-                const payload2 = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            toPayDetails: { 'formula':'balance + charge + recalculation + privilege + penalty', 'charge':'150.00', 'balance':'0.00', 'recalculation':null, 'privilege':null, 'penalty':null },
-                        },
-                    ],
-                }
-                const [data] = await registerBillingReceiptsByTestClient(admin, payload2)
-
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(1)
-                expect(billingReceipts).toHaveLength(1)
-                expect(data).toHaveLength(1)
-                expect(billingReceipts[0].id).toEqual(existingBillingReceiptId)
-                expect(billingReceipts[0].toPayDetails.charge).toEqual('150.00')
-            })
-
-            test('BillingReceipts are updated if only services are different', async () => {
-                const [organization] = await createTestOrganization(admin)
-                const [integration] = await createTestBillingIntegration(admin)
-                const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-                const receiptInput = createRegisterBillingReceiptsPayload()
-
-                const payload = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            services: undefined,
-                        },
-                    ],
-                }
-                const [ existingBillingReceipts ] = await registerBillingReceiptsByTestClient(admin, payload)
-                const existingBillingReceiptId = existingBillingReceipts[0].id
-
-                const payload2 = {
-                    context: { id: billingContext.id },
-                    receipts: [
-                        {
-                            ...receiptInput,
-                            services: [{ 'id':'928c97ef-5289-4daa-b80e-4b9fed50c629', 'name':'Квартплата', 'toPay':'100.00' }],
-                        },
-                    ],
-                }
-                const [data] = await registerBillingReceiptsByTestClient(admin, payload2)
-
-                const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-                const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-                expect(billingProperties).toHaveLength(1)
-                expect(billingReceipts).toHaveLength(1)
-                expect(data).toHaveLength(1)
-                expect(billingReceipts[0].id).toEqual(existingBillingReceiptId)
-                expect(billingReceipts[0].services[0].id).toEqual('928c97ef-5289-4daa-b80e-4b9fed50c629')
             })
         })
     })
 
-    describe('Real life cases', () => {
-        test('Management company loads 50 receipts for first time', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [],
-            }
-            for (let i = 0; i < 50; ++i) { payload.receipts.push(createRegisterBillingReceiptsPayload()) }
-
-            const [data] = await registerBillingReceiptsByTestClient(admin, payload)
-
-            const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-            expect(billingReceipts).toHaveLength(50)
-            expect(data).toHaveLength(50)
-        })
-
-        test('Management company changes billing recipient', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const tin1 = faker.random.alphaNumeric(8)
-            const routingNumber1 = faker.random.alphaNumeric(8)
-            const bankAccount1 = faker.random.alphaNumeric(8)
-
-            const receipt1 = createRegisterBillingReceiptsPayload({
-                tin: tin1,
-                routingNumber: routingNumber1,
-                bankAccount: bankAccount1,
+    describe('PeriodResolver',  () => {
+        test('should return an error on a wrong year format', async () => {
+            await catchErrorFrom(async () => {
+                await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [createJSONReceipt({ year: 23 })],
+                })
+            }, (e) => {
+                expect(e.errors[0].extensions.code).toEqual(ERRORS.WRONG_YEAR.code)
+                expect(e.errors[0].extensions.type).toEqual(ERRORS.WRONG_YEAR.type)
             })
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    receipt1,
-                ],
-            }
-
-            const [data] = await registerBillingReceiptsByTestClient(admin, payload)
-
-            const routingNumber2 = faker.random.alphaNumeric(8)
-            const bankAccount2 = faker.random.alphaNumeric(8)
-
-            const payload2 = {
-                context: { id: billingContext.id },
-                receipts: [
-                    { ...receipt1, ...{ routingNumber: routingNumber2, bankAccount: bankAccount2, importId: faker.random.alphaNumeric(8) } },
-                ],
-            }
-
-            const [data2] = await registerBillingReceiptsByTestClient(admin, payload2)
-            const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-            expect(data).toHaveLength(1)
-            expect(data2).toHaveLength(1)
-            expect(billingReceipts).toHaveLength(1)
-            expect(billingReceipts[0].recipient.bankAccount).toEqual(bankAccount2)
-            expect(billingReceipts[0].receiver.bankAccount).toEqual(bankAccount2)
         })
-
-        test('Management company signs up for management of new property', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const EXISTING_ADDRESS = 'Washington st, 4'
-            const NEW_ADDRESS = 'Washington st, 5'
-
-            const payload1 = {
-                context: { id: billingContext.id },
-                receipts: [],
-            }
-            for (let i = 0; i < 20; ++i) { payload1.receipts.push(createRegisterBillingReceiptsPayload({ address: EXISTING_ADDRESS })) }
-
-            const [data1] = await registerBillingReceiptsByTestClient(admin, payload1)
-
-            const payload2 = {
-                context: { id: billingContext.id },
-                receipts: [],
-            }
-            for (let i = 0; i < 20; ++i) { payload2.receipts.push(createRegisterBillingReceiptsPayload({ address: NEW_ADDRESS })) }
-
-            const [data2] = await registerBillingReceiptsByTestClient(admin, payload2)
-
-            const payload3 = {
-                context: { id: billingContext.id },
-                receipts: [],
-            }
-            for (let i = 0; i < 5; ++i) { payload3.receipts.push(createRegisterBillingReceiptsPayload({ address: EXISTING_ADDRESS })) }
-            for (let i = 5; i < 10; ++i) { payload3.receipts.push(createRegisterBillingReceiptsPayload({ address: NEW_ADDRESS })) }
-
-            const [data3] = await registerBillingReceiptsByTestClient(admin, payload3)
-
-            const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-            const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-            expect(data1).toHaveLength(20)
-            expect(data2).toHaveLength(20)
-            expect(data3).toHaveLength(10)
-            expect(billingReceipts).toHaveLength(50)
-            expect(billingProperties).toHaveLength(2)
-        })
-
-        test('Simple case is handled correctly', async () => {
-            const EXISTING_TEST_ADDRESS = 'TEST'
-            const EXISTING_TEST_UNIT_NAME = '0'
-            const EXISTING_TEST_ACCOUNT_NUMBER = '0'
-
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const [billingProperty] = await createTestBillingProperty(admin, billingContext, {
-                address: EXISTING_TEST_ADDRESS,
+        test('should return an error on a wrong month', async () => {
+            await catchErrorFrom(async () => {
+                await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [createJSONReceipt({ month: 13 })],
+                })
+            }, (e) => {
+                expect(e.errors[0].extensions.code).toEqual(ERRORS.WRONG_MONTH.code)
+                expect(e.errors[0].extensions.type).toEqual(ERRORS.WRONG_MONTH.type)
             })
-
-            await createTestBillingAccount(admin, billingContext, billingProperty, {
-                unitName: EXISTING_TEST_UNIT_NAME,
-                unitType: FLAT_UNIT_TYPE,
-                number: EXISTING_TEST_ACCOUNT_NUMBER,
-            })
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    createRegisterBillingReceiptsPayload({
-                        address: EXISTING_TEST_ADDRESS,
-
-                        unitType: FLAT_UNIT_TYPE,
-                        accountNumber: EXISTING_TEST_ACCOUNT_NUMBER,
-                        unitName: EXISTING_TEST_UNIT_NAME,
-                    }),
-                    createRegisterBillingReceiptsPayload({
-                        address: EXISTING_TEST_ADDRESS,
-                    }),
-                    createRegisterBillingReceiptsPayload(),
-                ],
-            }
-
-            const [ data ] = await registerBillingReceiptsByTestClient(admin, payload)
-            const billingProperties = await BillingProperty.getAll(admin, { context: { id: billingContext.id } })
-            const billingAccounts = await BillingAccount.getAll(admin, { context: { id: billingContext.id } })
-            const billingReceipts = await BillingReceipt.getAll(admin, { context: { id: billingContext.id } })
-
-            expect(billingProperties).toHaveLength(2)
-            expect(billingAccounts).toHaveLength(3)
-            expect(billingReceipts).toHaveLength(3)
-            expect(data).toHaveLength(3)
-        })
-
-        test('Hacker can not load receipts into other billing context related to other org', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-            const { managerUserClient } = await makeOrganizationIntegrationManager({ context: billingContext })
-
-            const [organization2] = await createTestOrganization(admin)
-            const [integration2] = await createTestBillingIntegration(admin)
-            const [billingContext2] = await createTestBillingIntegrationOrganizationContext(admin, organization2, integration2)
-            const { managerUserClient: hackerClient } = await makeOrganizationIntegrationManager({ context: billingContext2 })
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    createRegisterBillingReceiptsPayload(),
-                ],
-            }
-
-            await expectToThrowAccessDeniedErrorToResult(async () => {
-                await registerBillingReceiptsByTestClient(hackerClient, payload)
-            })
-
-            const billingProperties = await BillingProperty.getAll(managerUserClient, { context: { id: billingContext.id } })
-            const billingAccounts = await BillingAccount.getAll(managerUserClient, { context: { id: billingContext.id } })
-            const billingReceipts = await BillingReceipt.getAll(managerUserClient, { context: { id: billingContext.id } })
-
-            expect(billingProperties).toHaveLength(0)
-            expect(billingAccounts).toHaveLength(0)
-            expect(billingReceipts).toHaveLength(0)
         })
     })
 
-    describe('Validations', () => {
-        test('Mutation checks wrong context-id', async () => {
-
-            const payload = {
-                context: { id: admin.user.id },
-                receipts: [],
-            }
-
-            await expectToThrowGQLError(async () => {
-                await registerBillingReceiptsByTestClient(admin, payload)
-            }, { ...mutationErrors.BILLING_CONTEXT_NOT_FOUND }, 'result')
-        })
-
-        test('Mutation has limit on receipts', async () => {
-
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const RECEIPTS_LIMIT = 100
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [],
-            }
-
-            for (let i = 0; i < RECEIPTS_LIMIT + 1; ++i) {
-                payload.receipts.push(createRegisterBillingReceiptsPayload())
-            }
-
-            await expectToThrowGQLError(async () => {
-                await registerBillingReceiptsByTestClient(admin, payload)
-            }, { ...mutationErrors.RECEIPTS_LIMIT_HIT }, 'result')
-        })
-
-        test('Mutation checks wrong category-id', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    createRegisterBillingReceiptsPayload({
-                        category: { id: '928c97ef-5289-4daa-b80e-4b9fed50c630' }, // Wrong category id
-                    }),
-                ],
-            }
-
-            await catchErrorFrom(async () => {
-                await registerBillingReceiptsByTestClient(admin, payload)
-            }, (e) => {
-                expect(e.errors[0].message).toContain('BillingCategory is not found')
+    describe('RecipientResolver', () => {
+        test('Should create approved recipient on TIN matched', async () => {
+            const recipientInput = createRecipient({ tin: integration.organization.tin })
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt(recipientInput)],
             })
+            const recipient = await BillingRecipient.getOne(clients.admin, { bankAccount: recipientInput.bankAccount, context: { id: integration.billingContext.id } } )
+            expect(recipient.isApproved).toEqual(true)
         })
-
-        test('Mutation checks wrong period format (bad year)', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    createRegisterBillingReceiptsPayload({
-                        year: -200,
-                    }),
-                ],
-            }
-
-            await catchErrorFrom(async () => {
-                await registerBillingReceiptsByTestClient(admin, payload)
-            }, (e) => {
-                expect(e.errors[0].message).toContain('Year is wrong')
+        test('Should create not approved recipient on TIN  mismatch', async () => {
+            const recipientInput = createRecipient({ tin: integration.organization.tin + '0' })
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt(recipientInput)],
             })
+            const recipient = await BillingRecipient.getOne(clients.admin, { bankAccount: recipientInput.bankAccount, context: { id: integration.billingContext.id } } )
+            expect(recipient.isApproved).toEqual(false)
         })
-
-        test('Mutation checks wrong period format (bad month)', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    createRegisterBillingReceiptsPayload({
-                        month: 20,
-                    }),
-                ],
-            }
-
-            await catchErrorFrom(async () => {
-                await registerBillingReceiptsByTestClient(admin, payload)
-            }, (e) => {
-                console.log(e)
-                expect(e.errors[0].message).toContain('Month is wrong')
-                expect(e.errors[0].extensions.inputIndex).toEqual(0)
+        test('Should not modify isApproved if support user remove this flag', async () => {
+            const recipientInput = createRecipient({ bankAccount: faker.random.numeric(12), tin: integration.organization.tin  })
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt(recipientInput)],
             })
-        })
+            const recipient = await BillingRecipient.getOne(clients.admin, { bankAccount: recipientInput.bankAccount, context: { id: integration.billingContext.id } } )
+            expect(recipient.isApproved).toBeTruthy()
+            await BillingRecipient.update(clients.admin, recipient.id, { isApproved: false })
 
-        test('Mutation checks empty address', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    createRegisterBillingReceiptsPayload({
-                        address: '',
-                    }),
-                ],
-            }
-
-            await catchErrorFrom(async () => {
-                await registerBillingReceiptsByTestClient(admin, payload)
-            }, (e) => {
-                expect(e.errors[0].message).toContain('Address is empty')
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt(recipientInput)],
             })
+            const recipientAfterAnotherReceiptCreation = await BillingRecipient.getOne(clients.admin, { bankAccount: recipientInput.bankAccount, context: { id: integration.billingContext.id } } )
+            expect(recipientAfterAnotherReceiptCreation.isApproved).toBeFalsy()
         })
-
-        test('Errors are indexed correctly, mutation creates entities it can create, and sends errors for others', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    createRegisterBillingReceiptsPayload(),
-                    createRegisterBillingReceiptsPayload({
-                        month: 13,
-                    }),
-                ],
-            }
-            const [result, errors] = await registerBillingReceiptsByTestClient(admin, payload, { raw: true })
-
-            expect(result).toHaveLength(2)
-            expect(result[1]).toBeNull()
-
-            expect(errors).toHaveLength(1)
-            expect(errors[0].extensions.inputIndex).toEqual(1)
-            expect(errors[0].message).toContain('Month is wrong')
-        })
-
-        test('If there is an error, mutation creates entities it can create, and sends errors for others', async () => {
-            const [organization] = await createTestOrganization(admin)
-            const [integration] = await createTestBillingIntegration(admin)
-            const [billingContext] = await createTestBillingIntegrationOrganizationContext(admin, organization, integration)
-
-            const payload = {
-                context: { id: billingContext.id },
-                receipts: [
-                    createRegisterBillingReceiptsPayload(),
-                    createRegisterBillingReceiptsPayload(),
-                    createRegisterBillingReceiptsPayload(),
-                    createRegisterBillingReceiptsPayload({
-                        month: 13,
-                    }),
-                    createRegisterBillingReceiptsPayload({
-                        year: -20,
-                    }),
-                ],
-            }
-            const [result, errors] = await registerBillingReceiptsByTestClient(admin, payload, { raw: true })
-
-            expect(result).toHaveLength(5)
-            expect(result[3]).toBeNull()
-            expect(result[4]).toBeNull()
-            expect(errors).toHaveLength(2)
+        test('Should fill deprecated recipient field as we do not remove it still', async () => {
+            const recipientInput = createRecipient({ })
+            const [receipts] = await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt(recipientInput)],
+            })
+            const { recipient } = receipts[0]
+            expect(recipient).toHaveProperty('name')
+            expect(recipient).toHaveProperty('bankAccount')
+            expect(recipient).toHaveProperty('bic')
+            expect(recipient).toHaveProperty('tin')
         })
     })
+
+    describe('AccountResolver', () => {
+        test('Can exist 2 accounts with same number but different houses', async () => {
+            const accountNumber = randomNumber(10).toString()
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt({ accountNumber }), createJSONReceipt({ accountNumber })],
+            })
+            const accounts = await BillingAccount.getAll(clients.admin, { number: accountNumber, context: { id: integration.billingContext.id } } )
+            expect(accounts).toHaveLength(2)
+        })
+        test('Will set information from accountMeta to BillingAccount', async () => {
+            const accountMeta = {
+                globalId: createValidELS(),
+                importId: randomNumber(10).toString(),
+                fullName: faker.name.fullName(),
+                isClosed: true,
+            }
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt({ accountMeta })],
+            })
+            const account = await BillingAccount.getOne(clients.admin, { importId: accountMeta.importId, context: { id: integration.billingContext.id } } )
+            expect(account.fullName).toEqual(accountMeta.fullName)
+            expect(account.importId).toEqual(accountMeta.importId)
+            expect(account.globalId).toEqual(accountMeta.globalId)
+            expect(account.isClosed).toEqual(accountMeta.isClosed)
+            expect(account.ownerType).toEqual(BILLING_ACCOUNT_OWNER_TYPE_PERSON)
+            const companyAccountMeta = {
+                globalId: createValidELS() + '0',
+                importId: randomNumber(10).toString(),
+                fullName: faker.lorem.sentence(),
+                isClosed: false,
+            }
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt({ accountMeta: companyAccountMeta })],
+            })
+            const companyAccount = await BillingAccount.getOne(clients.admin, { importId: companyAccountMeta.importId, context: { id: integration.billingContext.id } } )
+            expect(companyAccount.fullName).toEqual(companyAccountMeta.fullName)
+            expect(companyAccount.importId).toEqual(companyAccountMeta.importId)
+            expect(companyAccount.globalId).toBeNull()
+            expect(companyAccount.isClosed).toEqual(companyAccountMeta.isClosed)
+            expect(companyAccount.ownerType).toEqual(BILLING_ACCOUNT_OWNER_TYPE_COMPANY)
+        })
+        test('It will not create another account on address change if importId is the same', async () => {
+            const importId = randomNumber(50).toString()
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt({ accountMeta: { importId } })],
+            })
+            const accounts = await BillingAccount.getAll(clients.admin, { importId, context: { id: integration.billingContext.id } } )
+            expect(accounts).toHaveLength(1)
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt({ accountMeta: { importId } })],
+            })
+            expect(accounts).toHaveLength(1)
+        })
+        test('isPerson validation "company" ownerType test', async () => {
+            const importId = randomNumber(50).toString()
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt({ accountMeta: { importId, fullName: 'ИП Фамилия' } })],
+            })
+            const accounts = await BillingAccount.getAll(clients.admin, { importId, context: { id: integration.billingContext.id } } )
+
+            expect(accounts[0].ownerType).toEqual(BILLING_ACCOUNT_OWNER_TYPE_COMPANY)
+        })
+        test('isPerson validation "person" ownerType test', async () => {
+            const importId = randomNumber(50).toString()
+            await registerBillingReceiptsByTestClient(clients.admin, {
+                context: { id: integration.billingContext.id },
+                receipts: [createJSONReceipt({ accountMeta: { importId, fullName: 'Фамилияоао Имя' } })],
+            })
+            const accounts = await BillingAccount.getAll(clients.admin, { importId, context: { id: integration.billingContext.id } } )
+
+            expect(accounts[0].ownerType).toEqual(BILLING_ACCOUNT_OWNER_TYPE_PERSON)
+        })
+    })
+
+    describe('ReceiptResolver', () => {
+        describe('Do not create new receipt on: ', () => {
+            test('[importId]. Address change (was resolved wrong and then fixed)', async () => {
+                const wrongAddress = createAddressWithUnit()
+                const receiptWithWrongAddress = createJSONReceipt({ address: wrongAddress })
+                await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [receiptWithWrongAddress],
+                })
+                const createdReceipt = await BillingReceipt.getOne(clients.admin, { importId: receiptWithWrongAddress.importId, context: { id: integration.billingContext.id } } )
+                const wrongHouseAddress = createdReceipt.property.address
+                const correctAddress = createAddressWithUnit()
+                const receiptWithCorrectAddress = { ...receiptWithWrongAddress, address: correctAddress }
+                await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [receiptWithCorrectAddress],
+                })
+                const updatedReceipt = await BillingReceipt.getOne(clients.admin, { importId: receiptWithCorrectAddress.importId, context: { id: integration.billingContext.id } } )
+                const correctHouseAddress = updatedReceipt.property.address
+                expect(updatedReceipt.property.address).not.toEqual(wrongHouseAddress)
+                expect(updatedReceipt.property.address).toEqual(correctHouseAddress)
+                expect(updatedReceipt.account.property.address).toEqual(correctHouseAddress)
+                expect(createdReceipt.account.id).toEqual(updatedReceipt.account.id)
+            })
+            test('[all] Category change (first load without services, then added services)', async () => {
+                const name = faker.lorem.sentence(3)
+                const serviceName = faker.lorem.sentence(2)
+                await createTestBillingCategory(clients.admin, { name, serviceNames: [serviceName] })
+                const createInput = createJSONReceipt({ importId: null, services: null })
+                const [[createdReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [createInput],
+                })
+                const [[updatedReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [{ ...createInput, services: [{ name: serviceName, toPay: '0' }] }],
+                })
+                expect(createdReceipt.category.id).not.toEqual(updatedReceipt.category.id)
+                expect(createdReceipt.id).toEqual(updatedReceipt.id)
+            })
+            test('[all] Recipient change do not creates new receipt', async () => {
+                const wrongRecipient = createRecipient()
+                const correctRecipient = createRecipient()
+                const createInput = createJSONReceipt({ importId: null, ...wrongRecipient })
+                const [[createdReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [createInput],
+                })
+                const [[updatedReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [{ ...createInput, ...correctRecipient }],
+                })
+                expect(createdReceipt.receiver.id).not.toEqual(updatedReceipt.receiver.id)
+                expect(createdReceipt.id).toEqual(updatedReceipt.id)
+            })
+            test('[all] Services change updates services in receipt', async () => {
+                const createServices = generateServicesData(3, '1000')
+                const createInput = createJSONReceipt({ services: createServices })
+                const [[createdReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [createInput],
+                })
+                const updateServices = generateServicesData(4, '1000')
+                const updateInput = createJSONReceipt({ ...createInput, services: updateServices })
+                const [[updatedReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [updateInput],
+                })
+                expect(createdReceipt.id).toEqual(updatedReceipt.id)
+                expect(createdReceipt.services).toHaveLength(3)
+                expect(updatedReceipt.services).toHaveLength(4)
+                expect(updatedReceipt.services.map(({ id, name, toPay }) => ({ id, name, toPay }) )).toEqual(updateServices.map(({ id, name, toPay }) => ({ id, name, toPay }) ))
+            })
+        })
+        describe('Common behaviour', () => {
+            test('Works on several receipts', async () => {
+                const receipt1 = createJSONReceipt()
+                const receipt2 = createJSONReceipt({
+                    ...receipt1,
+                    importId: null,
+                    address: createAddressWithUnit(),
+                    accountNumber: randomNumber(10).toString(),
+                })
+                const [createdReceipts] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [receipt1, receipt2],
+                })
+                expect(createdReceipts[0].id).not.toEqual(createdReceipts[1].id)
+            })
+            test('Fields with relations  are created', async () => {
+                const createInput = createJSONReceipt()
+                const [[createdReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [createInput],
+                })
+                expect(createdReceipt.property.id).toBeDefined()
+                expect(createdReceipt.account.id).toBeDefined()
+                expect(createdReceipt.receiver.id).toBeDefined()
+                expect(createdReceipt.context.id).toBeDefined()
+            })
+            test('Will not invoke update on nothing changed', async () => {
+                const createInput = createJSONReceipt()
+                const [[createdReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [createInput],
+                })
+                const [[updatedReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                    context: { id: integration.billingContext.id },
+                    receipts: [createInput],
+                })
+                expect(createdReceipt.id).toEqual(updatedReceipt.id)
+                expect(createdReceipt.v).toEqual(updatedReceipt.v)
+            })
+        })
+        describe('Resolvers tests', () => {
+            describe('receiptResolver', () => {
+                test('Update Receipt toPay field', async () => {
+                    const originalToPayValue = Big(faker.finance.amount(-100, 5000)).toFixed(2)
+                    const updatedToPayValue = Big(originalToPayValue).add(1000).toFixed(2)
+                    const createInput = createJSONReceipt({ toPay: originalToPayValue })
+                    const updateInput = { ...createInput, toPay: updatedToPayValue }
+                    const [[createdReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                        context: { id: integration.billingContext.id },
+                        receipts: [createInput],
+                    })
+                    const [[updatedReceipt]] = await registerBillingReceiptsByTestClient(clients.admin, {
+                        context: { id: integration.billingContext.id },
+                        receipts: [updateInput],
+                    })
+                    expect(createdReceipt.id).toEqual(updatedReceipt.id)
+                    expect(Big(updatedReceipt.toPay).toFixed(2)).toEqual(updatedToPayValue)
+                })
+            })
+
+        })
+    })
+
 })
